chore(components): reorganize gcpc llm components
PiperOrigin-RevId: 558581292
